% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_quality.R
\name{k_homogeneity}
\alias{k_homogeneity}
\title{Measures a degree of mutual dissimilarity between all objects in a cluster}
\usage{
k_homogeneity(my_k, df, sample_size)
}
\arguments{
\item{my_k}{\itemize{
\item a cluster number (from the k column)
}}

\item{df}{\itemize{
\item a tibble with the k column and the signature column
}}

\item{sample_size}{\itemize{
\item size of the sample (~maxhist)
}}
}
\description{
Measures a degree of mutual dissimilarity between all objects in a cluster
}
