% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_quality.R
\name{k_interdistance}
\alias{k_interdistance}
\title{It is an average distance between the focus cluster and all of the rest of the clusters}
\usage{
k_interdistance(my_k, df, sample_size)
}
\arguments{
\item{my_k}{\itemize{
\item a cluster number (from the k column)
}}

\item{df}{\itemize{
\item a tibble with the k column and the signature column
}}

\item{sample_size}{\itemize{
\item size of the sample (~maxhist)
}}
}
\description{
It is an average distance between the focus cluster and all of the rest of the clusters
}
