% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_extract2.R
\name{lsp_extract2}
\alias{lsp_extract2}
\title{Extracts a local landscape from categorical raster data based on its id and provided window.}
\usage{
lsp_extract2(id, windows_sf, x)
}
\arguments{
\item{id}{\itemize{
\item an id of a local landscpe
}}

\item{windows_sf}{\itemize{
\item the output of \link[motif:lsp_add_sf]{motif::lsp_add_sf}
}}

\item{x}{\itemize{
\item the \code{stars} object to extract from
}}
}
\description{
Extracts a local landscape from categorical raster data based on its id and provided window.
}
