% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_restructure.R
\name{lsp_restructure}
\alias{lsp_restructure}
\title{Converts a list-column signature into a number of columns}
\usage{
lsp_restructure(x)
}
\arguments{
\item{x}{\itemize{
\item an lsp object
}}
}
\description{
Converts a list-column signature into a number of columns
}
