% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifplus-package.R
\docType{package}
\name{motifplus-package}
\alias{motifplus}
\alias{motifplus-package}
\title{motifplus: Local Pattern Analysis Extended}
\description{
A package to test new ideas related to the motif package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Nowosad/motifplus}
  \item Report bugs at \url{https://github.com/Nowosad/motifplus/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

}
\keyword{internal}
